+++ src/connectedmachine/azext_connectedmachine/vendored_sdks/connectedmachine/operations/_private_link_scopes_operations.py
@@ -64,7 +64,7 @@ class PrivateLinkScopesOperations(object):
             401: ClientAuthenticationError, 404: ResourceNotFoundError, 409: ResourceExistsError
         }
         error_map.update(kwargs.pop('error_map', {}))
-        api_version = "2022-03-10"
+        api_version = "2022-12-27"
         accept = "application/json"
 
         def prepare_request(next_link=None):
@@ -135,7 +135,7 @@ class PrivateLinkScopesOperations(object):
             401: ClientAuthenticationError, 404: ResourceNotFoundError, 409: ResourceExistsError
         }
         error_map.update(kwargs.pop('error_map', {}))
-        api_version = "2022-03-10"
+        api_version = "2022-12-27"
         accept = "application/json"
 
         def prepare_request(next_link=None):
@@ -199,7 +199,7 @@ class PrivateLinkScopesOperations(object):
             401: ClientAuthenticationError, 404: ResourceNotFoundError, 409: ResourceExistsError
         }
         error_map.update(kwargs.pop('error_map', {}))
-        api_version = "2022-03-10"
+        api_version = "2022-12-27"
         accept = "application/json"
 
         # Construct URL
@@ -321,7 +321,7 @@ class PrivateLinkScopesOperations(object):
             401: ClientAuthenticationError, 404: ResourceNotFoundError, 409: ResourceExistsError
         }
         error_map.update(kwargs.pop('error_map', {}))
-        api_version = "2022-03-10"
+        api_version = "2022-12-27"
         accept = "application/json"
 
         # Construct URL
@@ -386,7 +386,7 @@ class PrivateLinkScopesOperations(object):
             401: ClientAuthenticationError, 404: ResourceNotFoundError, 409: ResourceExistsError
         }
         error_map.update(kwargs.pop('error_map', {}))
-        api_version = "2022-03-10"
+        api_version = "2022-12-27"
         content_type = kwargs.pop("content_type", "application/json")
         accept = "application/json"
 
@@ -440,8 +440,7 @@ class PrivateLinkScopesOperations(object):
         **kwargs  # type: Any
     ):
         # type: (...) -> "models.HybridComputePrivateLinkScope"
-        """Updates an existing PrivateLinkScope's tags. To update other fields use the CreateOrUpdate
-        method.
+        """Updates an existing PrivateLinkScope's tags.
 
         :param resource_group_name: The name of the resource group. The name is case insensitive.
         :type resource_group_name: str
@@ -460,7 +459,7 @@ class PrivateLinkScopesOperations(object):
             401: ClientAuthenticationError, 404: ResourceNotFoundError, 409: ResourceExistsError
         }
         error_map.update(kwargs.pop('error_map', {}))
-        api_version = "2022-03-10"
+        api_version = "2022-12-27"
         content_type = kwargs.pop("content_type", "application/json")
         accept = "application/json"
 
@@ -525,7 +524,7 @@ class PrivateLinkScopesOperations(object):
             401: ClientAuthenticationError, 404: ResourceNotFoundError, 409: ResourceExistsError
         }
         error_map.update(kwargs.pop('error_map', {}))
-        api_version = "2022-03-10"
+        api_version = "2022-12-27"
         accept = "application/json"
 
         # Construct URL
@@ -586,7 +585,7 @@ class PrivateLinkScopesOperations(object):
             401: ClientAuthenticationError, 404: ResourceNotFoundError, 409: ResourceExistsError
         }
         error_map.update(kwargs.pop('error_map', {}))
-        api_version = "2022-03-10"
+        api_version = "2022-12-27"
         accept = "application/json"
 
         # Construct URL