+++ src/containerapp/azext_containerapp/containerapp_job_decorator.py
@@ -19,11 +19,12 @@ from msrestazure.tools import parse_resource_id, is_valid_resource_id
 from msrest.exceptions import DeserializationError
 
 from ._decorator_utils import process_loaded_yaml, load_yaml_file, create_deserializer
-from ._constants import HELLO_WORLD_IMAGE, CONTAINER_APPS_RP
+from ._constants import HELLO_WORLD_IMAGE, CONTAINER_APPS_RP, CONNECTED_ENVIRONMENT_RESOURCE_TYPE, \
+    MANAGED_ENVIRONMENT_TYPE, CONNECTED_ENVIRONMENT_TYPE
 from ._validators import validate_create
 from .base_resource import BaseResource
-from ._clients import ManagedEnvironmentClient
-from ._client_factory import handle_raw_exception
+from ._clients import ManagedEnvironmentClient, ConnectedEnvironmentClient, ManagedEnvironmentPreviewClient
+from ._client_factory import handle_raw_exception, handle_non_404_exception
 
 from ._models import (
     JobConfiguration as JobConfigurationModel,
@@ -405,7 +406,7 @@ class ContainerAppJobCreateDecorator(ContainerAppJobDecorator):
                 set_managed_identity(self.cmd, self.get_argument_resource_group_name(), self.containerappjob_def, user_assigned=[self.get_argument_registry_identity()])
 
     def set_up_create_containerapp_job_yaml(self, name, file_name):
-        if self.get_argument_image() or self.get_argument_managed_env() or self.get_argument_trigger_type() or self.get_argument_replica_timeout() or self.get_argument_replica_retry_limit() or \
+        if self.get_argument_image() or self.get_argument_trigger_type() or self.get_argument_replica_timeout() or self.get_argument_replica_retry_limit() or \
                 self.get_argument_replica_completion_count() or self.get_argument_parallelism() or self.get_argument_cron_expression() or self.get_argument_cpu() or self.get_argument_memory() or self.get_argument_registry_server() or \
                 self.get_argument_registry_user() or self.get_argument_registry_pass() or self.get_argument_secrets() or self.get_argument_env_vars() or \
                 self.get_argument_startup_command() or self.get_argument_args() or self.get_argument_tags():
@@ -460,14 +461,18 @@ class ContainerAppJobCreateDecorator(ContainerAppJobDecorator):
             del self.containerappjob_def["workloadProfileName"]
 
         # Validate managed environment
+        env_id = self.containerappjob_def["properties"]['environmentId']
+        env_info = None
+        if self.get_argument_managed_env():
+            if not self.get_argument_disable_warnings() and env_id is not None and env_id != self.get_argument_managed_env():
+                logger.warning('The environmentId was passed along with --yaml. The value entered with --environment will be ignored, and the configuration defined in the yaml will be used instead')
+            if env_id is None:
+                env_id = self.get_argument_managed_env()
+                safe_set(self.containerappjob_def, "properties", "environmentId", value=env_id)
+
         if not self.containerappjob_def["properties"].get('environmentId'):
             raise RequiredArgumentMissingError(
                 'environmentId is required. This can be retrieved using the `az containerapp env show -g MyResourceGroup -n MyContainerappEnvironment --query id` command. Please see https://aka.ms/azure-container-apps-yaml for a valid containerapps YAML spec.')
-
-        env_id = self.containerappjob_def["properties"]['environmentId']
-        env_name = None
-        env_rg = None
-        env_info = None
 
         if is_valid_resource_id(env_id):
             parsed_managed_env = parse_resource_id(env_id)
@@ -478,8 +483,8 @@ class ContainerAppJobCreateDecorator(ContainerAppJobDecorator):
 
         try:
             env_info = self.get_environment_client().show(cmd=self.cmd, resource_group_name=env_rg, name=env_name)
-        except:
-            pass
+        except Exception as e:
+            handle_non_404_exception(e)
 
         if not env_info:
             raise ValidationError("The environment '{}' in resource group '{}' was not found".format(env_name, env_rg))
@@ -487,3 +492,58 @@ class ContainerAppJobCreateDecorator(ContainerAppJobDecorator):
         # Validate location
         if not self.containerappjob_def.get('location'):
             self.containerappjob_def['location'] = env_info['location']
+
+
+class ContainerAppJobPreviewCreateDecorator(ContainerAppJobCreateDecorator):
+    def construct_payload(self):
+        super().construct_payload()
+        self.set_up_extended_location()
+
+    def set_up_extended_location(self):
+        if self.get_argument_environment_type() == CONNECTED_ENVIRONMENT_TYPE:
+            if not self.containerappjob_def.get('extendedLocation'):
+                env_id = safe_get(self.containerappjob_def, "properties", 'environmentId') or self.get_argument_managed_env()
+                parsed_env = parse_resource_id(env_id)
+                env_name = parsed_env['name']
+                env_rg = parsed_env['resource_group']
+                env_info = self.get_environment_client().show(cmd=self.cmd, resource_group_name=env_rg, name=env_name)
+                self.containerappjob_def["extendedLocation"] = env_info["extendedLocation"]
+
+    def get_environment_client(self):
+        if self.get_argument_yaml():
+            env = safe_get(self.containerappjob_def, "properties", "environmentId")
+        else:
+            env = self.get_argument_managed_env()
+
+        environment_type = self.get_argument_environment_type()
+        if not env and not environment_type:
+            return ManagedEnvironmentClient
+
+        parsed_env = parse_resource_id(env)
+
+        # Validate environment type
+        if parsed_env.get('resource_type').lower() == CONNECTED_ENVIRONMENT_RESOURCE_TYPE.lower():
+            if environment_type == MANAGED_ENVIRONMENT_TYPE:
+                logger.warning(f"User passed a connectedEnvironment resource id but did not specify --environment-type {CONNECTED_ENVIRONMENT_TYPE}. Using environment type {CONNECTED_ENVIRONMENT_TYPE}.")
+            environment_type = CONNECTED_ENVIRONMENT_TYPE
+        else:
+            if environment_type == CONNECTED_ENVIRONMENT_TYPE:
+                logger.warning(f"User passed a managedEnvironment resource id but specified --environment-type {CONNECTED_ENVIRONMENT_TYPE}. Using environment type {MANAGED_ENVIRONMENT_TYPE}.")
+            environment_type = MANAGED_ENVIRONMENT_TYPE
+
+        self.set_argument_environment_type(environment_type)
+        self.set_argument_managed_env(env)
+
+        if environment_type == CONNECTED_ENVIRONMENT_TYPE:
+            return ConnectedEnvironmentClient
+        else:
+            return ManagedEnvironmentPreviewClient
+
+    def get_argument_environment_type(self):
+        return self.get_param("environment_type")
+
+    def set_argument_managed_env(self, managed_env):
+        self.set_param("managed_env", managed_env)
+
+    def set_argument_environment_type(self, environment_type):
+        self.set_param("environment_type", environment_type)