+++ src/networkcloud/azext_networkcloud/tests/latest/test_storageappliance.py
@@ -9,8 +9,9 @@
 StorageAppliance tests scenarios
 """
 
+from azure.cli.testsdk import ResourceGroupPreparer, ScenarioTest
 from azure.cli.testsdk.scenario_tests import AllowLargeResponse
-from azure.cli.testsdk import ScenarioTest, ResourceGroupPreparer
+
 from .config import CONFIG