+++ src/spring/azext_spring/custom.py
@@ -205,6 +205,15 @@ def spring_stop(cmd, client, resource_group, name, no_wait=False):
     return sdk_no_wait(no_wait, client.services.begin_stop, resource_group_name=resource_group, service_name=name)
 
 
+def spring_flush_vnet_dns_setting(cmd, client, resource_group, name, no_wait=False):
+    resource = client.services.get(resource_group, name)
+    state = resource.properties.provisioning_state
+    power_state = resource.properties.power_state
+    if state != "Succeeded" or power_state != "Running":
+        raise ClientRequestError("Service is in Provisioning State({}) and Power State({}), flush vnet dns setting cannot be performed.".format(state, power_state))
+    return sdk_no_wait(no_wait, client.services.begin_flush_vnet_dns_setting, resource_group_name=resource_group, service_name=name)
+
+
 def spring_list(cmd, client, resource_group=None):
     if resource_group is None:
         return client.services.list_by_subscription()
@@ -1243,8 +1252,14 @@ def _get_app_log(url, auth, format_json, exceptions, chunk_size=None, stderr=Fal
     with requests.get(url, stream=True, auth=auth) as response:
         try:
             if response.status_code != 200:
+                failure_reason = response.reason
+                if response.content:
+                    if isinstance(response.content, bytes):
+                        failure_reason = "{}:{}".format(failure_reason, response.content.decode('utf-8'))
+                    else:
+                        failure_reason = "{}:{}".format(failure_reason, response.content)
                 raise CLIError("Failed to connect to the server with status code '{}' and reason '{}'".format(
-                    response.status_code, response.reason))
+                    response.status_code, failure_reason))
             std_encoding = sys.stdout.encoding
 
             formatter = build_formatter()