+++ src/spring/HISTORY.md
@@ -1,5 +1,13 @@
 Release History
 ===============
+1.14.3
+---
+* Make error message for `az spring app logs` more readable.  
+
+1.14.2
+---
+* Add new command `az spring flush-virtualnetwork-dns-settings` to flush virtual network DNS settings for the service instance.
+
 1.14.1
 ---
 * Support up to 1000 app instances in Enterprise tier.